Total update of IRequests hierarchy, added new requests, and removed unnecesary #include into IRequests
