+++ IRequests/Chat/RequestAddUserToTheChat.cpp
@@ -1,24 +1,20 @@
 #pragma once
-#include "IRequests.h"
-#include "RequestLeaveChat.h"
-#include <cpprest/json.h>
-#include "../StringUtils.h"
-using namespace web;
+#include "RequestAddUserToTheChat.h"
 
-RequestLeaveChat::RequestLeaveChat(IDatabase* db, const std::string& userAccessToken, const unsigned long& chatId, const std::string& userLogin) : IRequests(db),
+
+RequestAddUserToTheChat::RequestAddUserToTheChat(IDatabase* db, const std::string& userAccessToken, const unsigned long& chatId, const std::string& userLogin) : IRequests(db),
 user_access_token(userAccessToken),
 chat_id(chatId),user_login(userLogin) {}
 
-void RequestLeaveChat::DoRequest() {
+void RequestAddUserToTheChat::DoRequest() {
     json::value result;
     try {
-        if (this->db->RemoveUserFromChat(this->user_access_token, this->user_login,this->chat_id)) {
+        if (this->db->AddUserToChat(this->user_access_token, this->user_login,this->chat_id)) {
             result[L"status"] = json::value::string(L"OK");
-
-			  this->answercontainer->SetStatusCode(status_codes::OK);
-		}
+            this->answercontainer->SetStatusCode(status_codes::OK);
+        }
         else {
-            result[L"what"] = json::value::string(to_wstring("Cannot leave the chat"));
+            result[L"what"] = json::value::string(to_wstring("Cannot join to this chat"));
             this->answercontainer->SetStatusCode(status_codes::Forbidden);
         }
     }
