+++ IRequests/Chat/RequestAddUserToTheChat.h
@@ -1,6 +1,5 @@
 #pragma once
-#include <cpprest/json.h>
-using namespace web;
+#include "../IRequests.h"
 
 class RequestAddUserToTheChat : public IRequests {
 private:
