--- IRequests/RequestGetMessages.h
@@ -1,13 +1,0 @@
-#pragma once
-#include <cpprest/json.h>
-using namespace web;
-
-class RequestGetMessages : public IRequests {
-private:
-    std::string user_access_token;
-    unsigned long chat_id, last_message_id;
-public:
-    RequestGetMessages(IDatabase* db, const std::string& userAccessToken, const unsigned long& chatId, const unsigned long& lastMessageId);
-
-    void DoRequest();
-};
