+++ IRequests/User/RequestSignUp.h
@@ -1,0 +1,11 @@
+#pragma once
+#include "../IRequests.h"
+
+class RequestSignUp : public IRequests {
+private:
+    ISXModel::User user;
+public:
+    RequestSignUp(IDatabase* db, const ISXModel::User& user);
+    void DoRequest();
+};
+
